pull logic out from react component
