+++ packages/@uppy/google-photos-picker/src/GooglePhotosPicker.tsx
@@ -7,11 +7,8 @@ import {
   tokenStorage,
 } from '@uppy/companion-client'
 
+import type { PickedItem } from '@uppy/provider-views/lib/GooglePicker/googlePicker.js'
 import type { Body, Meta } from '@uppy/utils/lib/UppyFile'
-import {
-  type PickedItem,
-  type PluginState,
-} from '@uppy/provider-views/lib/GooglePicker/GooglePickerView.js'
 import type { AsyncStore, BaseProviderPlugin } from '@uppy/core/lib/Uppy.js'
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -59,7 +56,7 @@ export default class GooglePhotosPicker<
     M extends Meta & { width: number; height: number },
     B extends Body,
   >
-  extends UIPlugin<GooglePhotosPickerOptions, M, B, PluginState>
+  extends UIPlugin<GooglePhotosPickerOptions, M, B>
   implements BaseProviderPlugin
 {
   static VERSION = packageJson.version
@@ -137,7 +134,6 @@ export default class GooglePhotosPicker<
     <GooglePickerView
       storage={this.storage}
       pickerType="photos"
-      plugin={this}
       uppy={this.uppy}
       clientId={this.opts.clientId}
       onFilesPicked={this.handleFilesPicked}
